Safari-611.1.10/LayoutTests/accessibility/mac/slider-thumb-value-crash-expected.txt

This tests that getting the indicator's value from a removed slider won't cause crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS indicator.intValue is 0
PASS successfullyParsed is true

TEST COMPLETE
